requirements.txt
discord
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pillow
